+++ internal/models/eventParticipants.go
@@ -24,6 +24,7 @@ type EventParticipants struct {
 	EventParticipantsCreatedAt             primitive.DateTime          `bson:"event_participants_created_at,omitempty" json:"event_participants_created_at"`
 	EventParticipantsUserDetail            *UsersAgg                   `bson:"event_participants_user_detail,omitempty" json:"event_participants_user_detail,omitempty"`
 	EventParticipantsInvitation            *EventParticipantInvitation `bson:"event_participant_invitation,omitempty" json:"event_participant_invitation,omitempty"`
+	EventParticipantsInvitedBy             *UsersInvitation            `bson:"event_participants_invited_by,omitempty" json:"event_participants_invited_by,omitempty"`
 }
 type EventParticipantInvitation struct {
 	InvitationMessage  string `bson:"invitation_message,omitempty" json:"invitation_message"`